🚀 Merge 'develop' branch into 'main' branch

+++ src/main/java/site/haruhana/www/dto/submission/response/SubmissionResponseDto.java
@@ -3,7 +3,6 @@ package site.haruhana.www.dto.submission.response;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Builder;
 import lombok.Getter;
-import site.haruhana.www.entity.problem.ProblemType;
 import site.haruhana.www.entity.submission.Submission;
 
 import java.time.LocalDateTime;
@@ -12,19 +11,19 @@ import java.time.LocalDateTime;
 @Builder
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class SubmissionResponseDto {
-    
+
     private Long id;
     private LocalDateTime submittedAt;
 
     // 객관식 문제 전용 필드
     private Boolean isCorrect;
-    
+
     // 주관식 문제 전용 필드
-    private Integer score;
+    private Double score;
     private String feedback;
     private LocalDateTime feedbackProvidedAt;
     private Boolean isPending;  // 채점 대기중 여부
-    
+
     /**
      * 객관식 문제 제출에 대한 응답 DTO 생성하는 정적 팩토리 메서드
      */
@@ -35,13 +34,13 @@ public class SubmissionResponseDto {
                 .isCorrect(submission.getIsCorrect())
                 .build();
     }
-    
+
     /**
      * 주관식 문제 제출에 대한 응답 DTO 생성하는 정적 팩토리 메서드
      */
     public static SubmissionResponseDto fromSubjective(Submission submission) {
         boolean isPending = submission.getFeedback() == null;
-        
+
         return SubmissionResponseDto.builder()
                 .id(submission.getId())
                 .submittedAt(submission.getSubmittedAt())
